% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_tidy.R
\name{cohort_tidy}
\alias{cohort_tidy}
\title{Cohort Tidy}
\usage{
cohort_tidy(data, ..., name = "X1")
}
\arguments{
\item{data}{a `data.frame` or `tibble` containing data of interest}

\item{...}{column name assignments to be passed to `transmute`}

\item{name}{name for new nested observation set column--  a single character value}
}
\value{
A nested tidycohort tibble/ dataframe
}
\description{
Transform observational data into tidycohort format
}
\examples{
\dontrun{
cohort_tidy(iris, id=Species, value_numeric=Sepal.Length, name="sepal_obs")
}
}
