% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_to_obs.R
\name{filter_obs}
\alias{filter_obs}
\title{Filter Observation Sets}
\usage{
filter_obs(data, formula, x)
}
\arguments{
\item{data}{a tidycohort}

\item{formula}{a formula of the form `new_column_name ~ target_column`}

\item{x}{the filter to be applied}
}
\value{
A tidycohort with a new column
}
\description{
Create a new observation set from another with an arbitrary filter
}
\examples{
\dontrun{filter_obs(X, obs_2 ~ observation, obs_type == 440922)}
}
