% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_to_vector.R
\name{obs_to_scalar}
\alias{obs_to_scalar}
\alias{summarise_obs}
\alias{any_obs_of_type}
\alias{any_obs}
\alias{index_time}
\title{Observation Sets to Vectors}
\usage{
summarise_obs(data, formula, summary)

any_obs_of_type(data, formula, type)

any_obs(data, formula)

index_time(data, formula)
}
\arguments{
\item{data}{a tidycohort data frame}

\item{formula}{a formula of the form \code{new_column_name ~ target_column}}

\item{summary}{the summary statement to be used}

\item{type}{the observation type, a typical `obs_type` in the observation set}
}
\value{
A tidycohort with a new column determined by the LHS of \code{forumla}
}
\description{
Summarize an observation set (nested data frames) to a single value for each patient
}
\details{
`any_obs_of_type()` checks for observations of type `obs_type` in an observation set

`any_obs()` checks for any observations at all in an observation set

`index_time()` extracts the earliest `start_time` in an observation set
}
\examples{
\dontrun{summarise_obs(X, new_col ~ old_col, sum(obs_type == "xyz"))}
\dontrun{any_obs(X, any_diagnoses_obs ~diagnoses)}
\dontrun{index_time(Y, diabetes_index_time ~ diabetes)}
}
